Manipulating Dates

date_shift adds days, weeks, months, quarters or years to dates. date_month, date_quarter and date_year return the first day of the period and are useful for customized aggregation of data frames. For standard aggregation, use ts_frequency().

date_shift(x, by = NULL)

date_month(x)

date_quarter(x)

date_year(x)

Arguments

x
Date or POSIXct. If POSIXct, it is converted into Date.

by
passed on to base::seq.Date(). Argument can be specified in several ways: - A number, taken to be in days. - A object of class difftime - A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".

Value

an object of class Date

See also

ts_frequency() for standard aggregation. date_shift(), for shifting time stamps of a ts-boxable object.

Examples

date_month(ts_df(AirPassengers)$time)

#>   [1] "1949-01-01" "1949-02-01" "1949-03-01" "1949-04-01" "1949-05-01"
#>   [6] "1949-06-01" "1949-07-01" "1949-08-01" "1949-09-01" "1949-10-01"
#>  [11] "1949-11-01" "1949-12-01" "1950-01-01" "1950-02-01" "1950-03-01"
#>  [16] "1950-04-01" "1950-05-01" "1950-06-01" "1950-07-01" "1950-08-01"
#>  [21] "1950-09-01" "1950-10-01" "1950-11-01" "1950-12-01" "1951-01-01"
#>  [26] "1951-02-01" "1951-03-01" "1951-04-01" "1951-05-01" "1951-06-01"
#>  [31] "1951-07-01" "1951-08-01" "1951-09-01" "1951-10-01" "1951-11-01"
#>  [36] "1951-12-01" "1952-01-01" "1952-02-01" "1952-03-01" "1952-04-01"
#>  [41] "1952-05-01" "1952-06-01" "1952-07-01" "1952-08-01" "1952-09-01"
#>  [46] "1952-10-01" "1952-11-01" "1952-12-01" "1953-01-01" "1953-02-01"
#>  [51] "1953-03-01" "1953-04-01" "1953-05-01" "1953-06-01" "1953-07-01"
#>  [56] "1953-08-01" "1953-09-01" "1953-10-01" "1953-11-01" "1953-12-01"
#>  [61] "1954-01-01" "1954-02-01" "1954-03-01" "1954-04-01" "1954-05-01"
#>  [66] "1954-06-01" "1954-07-01" "1954-08-01" "1954-09-01" "1954-10-01"
#>  [71] "1954-11-01" "1954-12-01" "1955-01-01" "1955-02-01" "1955-03-01"
#>  [76] "1955-04-01" "1955-05-01" "1955-06-01" "1955-07-01" "1955-08-01"
#>  [81] "1955-09-01" "1955-10-01" "1955-11-01" "1955-12-01" "1956-01-01"
#>  [86] "1956-02-01" "1956-03-01" "1956-04-01" "1956-05-01" "1956-06-01"
#>  [91] "1956-07-01" "1956-08-01" "1956-09-01" "1956-10-01" "1956-11-01"
#>  [96] "1956-12-01" "1957-01-01" "1957-02-01" "1957-03-01" "1957-04-01"
#> [101] "1957-05-01" "1957-06-01" "1957-07-01" "1957-08-01" "1957-09-01"
#> [106] "1957-10-01" "1957-11-01" "1957-12-01" "1958-01-01" "1958-02-01"
#> [111] "1958-03-01" "1958-04-01" "1958-05-01" "1958-06-01" "1958-07-01"
#> [116] "1958-08-01" "1958-09-01" "1958-10-01" "1958-11-01" "1958-12-01"
#> [121] "1959-01-01" "1959-02-01" "1959-03-01" "1959-04-01" "1959-05-01"
#> [126] "1959-06-01" "1959-07-01" "1959-08-01" "1959-09-01" "1959-10-01"
#> [131] "1959-11-01" "1959-12-01" "1960-01-01" "1960-02-01" "1960-03-01"
#> [136] "1960-04-01" "1960-05-01" "1960-06-01" "1960-07-01" "1960-08-01"
#> [141] "1960-09-01" "1960-10-01" "1960-11-01" "1960-12-01"

date_year(ts_df(AirPassengers)$time)

#>   [1] "1949-01-01" "1949-01-01" "1949-01-01" "1949-01-01" "1949-01-01"
#>   [6] "1949-01-01" "1949-01-01" "1949-01-01" "1949-01-01" "1949-01-01"
#>  [11] "1949-01-01" "1949-01-01" "1950-01-01" "1950-01-01" "1950-01-01"
#>  [16] "1950-01-01" "1950-01-01" "1950-01-01" "1950-01-01" "1950-01-01"
#>  [21] "1950-01-01" "1950-01-01" "1950-01-01" "1950-01-01" "1951-01-01"
#>  [26] "1951-01-01" "1951-01-01" "1951-01-01" "1951-01-01" "1951-01-01"
#>  [31] "1951-01-01" "1951-01-01" "1951-01-01" "1951-01-01" "1951-01-01"
#>  [36] "1951-01-01" "1952-01-01" "1952-01-01" "1952-01-01" "1952-01-01"
#>  [41] "1952-01-01" "1952-01-01" "1952-01-01" "1952-01-01" "1952-01-01"
#>  [46] "1952-01-01" "1952-01-01" "1952-01-01" "1953-01-01" "1953-01-01"
#>  [51] "1953-01-01" "1953-01-01" "1953-01-01" "1953-01-01" "1953-01-01"
#>  [56] "1953-01-01" "1953-01-01" "1953-01-01" "1953-01-01" "1953-01-01"
#>  [61] "1954-01-01" "1954-01-01" "1954-01-01" "1954-01-01" "1954-01-01"
#>  [66] "1954-01-01" "1954-01-01" "1954-01-01" "1954-01-01" "1954-01-01"
#>  [71] "1954-01-01" "1954-01-01" "1955-01-01" "1955-01-01" "1955-01-01"
#>  [76] "1955-01-01" "1955-01-01" "1955-01-01" "1955-01-01" "1955-01-01"
#>  [81] "1955-01-01" "1955-01-01" "1955-01-01" "1955-01-01" "1956-01-01"
#>  [86] "1956-01-01" "1956-01-01" "1956-01-01" "1956-01-01" "1956-01-01"
#>  [91] "1956-01-01" "1956-01-01" "1956-01-01" "1956-01-01" "1956-01-01"
#>  [96] "1956-01-01" "1957-01-01" "1957-01-01" "1957-01-01" "1957-01-01"
#> [101] "1957-01-01" "1957-01-01" "1957-01-01" "1957-01-01" "1957-01-01"
#> [106] "1957-01-01" "1957-01-01" "1957-01-01" "1958-01-01" "1958-01-01"
#> [111] "1958-01-01" "1958-01-01" "1958-01-01" "1958-01-01" "1958-01-01"
#> [116] "1958-01-01" "1958-01-01" "1958-01-01" "1958-01-01" "1958-01-01"
#> [121] "1959-01-01" "1959-01-01" "1959-01-01" "1959-01-01" "1959-01-01"
#> [126] "1959-01-01" "1959-01-01" "1959-01-01" "1959-01-01" "1959-01-01"
#> [131] "1959-01-01" "1959-01-01" "1960-01-01" "1960-01-01" "1960-01-01"
#> [136] "1960-01-01" "1960-01-01" "1960-01-01" "1960-01-01" "1960-01-01"
#> [141] "1960-01-01" "1960-01-01" "1960-01-01" "1960-01-01"

date_shift(ts_df(AirPassengers)$time, 14)

#>   [1] "1949-01-15" "1949-02-15" "1949-03-15" "1949-04-15" "1949-05-15"
#>   [6] "1949-06-15" "1949-07-15" "1949-08-15" "1949-09-15" "1949-10-15"
#>  [11] "1949-11-15" "1949-12-15" "1950-01-15" "1950-02-15" "1950-03-15"
#>  [16] "1950-04-15" "1950-05-15" "1950-06-15" "1950-07-15" "1950-08-15"
#>  [21] "1950-09-15" "1950-10-15" "1950-11-15" "1950-12-15" "1951-01-15"
#>  [26] "1951-02-15" "1951-03-15" "1951-04-15" "1951-05-15" "1951-06-15"
#>  [31] "1951-07-15" "1951-08-15" "1951-09-15" "1951-10-15" "1951-11-15"
#>  [36] "1951-12-15" "1952-01-15" "1952-02-15" "1952-03-15" "1952-04-15"
#>  [41] "1952-05-15" "1952-06-15" "1952-07-15" "1952-08-15" "1952-09-15"
#>  [46] "1952-10-15" "1952-11-15" "1952-12-15" "1953-01-15" "1953-02-15"
#>  [51] "1953-03-15" "1953-04-15" "1953-05-15" "1953-06-15" "1953-07-15"
#>  [56] "1953-08-15" "1953-09-15" "1953-10-15" "1953-11-15" "1953-12-15"
#>  [61] "1954-01-15" "1954-02-15" "1954-03-15" "1954-04-15" "1954-05-15"
#>  [66] "1954-06-15" "1954-07-15" "1954-08-15" "1954-09-15" "1954-10-15"
#>  [71] "1954-11-15" "1954-12-15" "1955-01-15" "1955-02-15" "1955-03-15"
#>  [76] "1955-04-15" "1955-05-15" "1955-06-15" "1955-07-15" "1955-08-15"
#>  [81] "1955-09-15" "1955-10-15" "1955-11-15" "1955-12-15" "1956-01-15"
#>  [86] "1956-02-15" "1956-03-15" "1956-04-15" "1956-05-15" "1956-06-15"
#>  [91] "1956-07-15" "1956-08-15" "1956-09-15" "1956-10-15" "1956-11-15"
#>  [96] "1956-12-15" "1957-01-15" "1957-02-15" "1957-03-15" "1957-04-15"
#> [101] "1957-05-15" "1957-06-15" "1957-07-15" "1957-08-15" "1957-09-15"
#> [106] "1957-10-15" "1957-11-15" "1957-12-15" "1958-01-15" "1958-02-15"
#> [111] "1958-03-15" "1958-04-15" "1958-05-15" "1958-06-15" "1958-07-15"
#> [116] "1958-08-15" "1958-09-15" "1958-10-15" "1958-11-15" "1958-12-15"
#> [121] "1959-01-15" "1959-02-15" "1959-03-15" "1959-04-15" "1959-05-15"
#> [126] "1959-06-15" "1959-07-15" "1959-08-15" "1959-09-15" "1959-10-15"
#> [131] "1959-11-15" "1959-12-15" "1960-01-15" "1960-02-15" "1960-03-15"
#> [136] "1960-04-15" "1960-05-15" "1960-06-15" "1960-07-15" "1960-08-15"
#> [141] "1960-09-15" "1960-10-15" "1960-11-15" "1960-12-15"

date_shift(ts_df(AirPassengers)$time, "7 week")

#>   [1] "1949-02-19" "1949-03-22" "1949-04-19" "1949-05-20" "1949-06-19"
#>   [6] "1949-07-20" "1949-08-19" "1949-09-19" "1949-10-20" "1949-11-19"
#>  [11] "1949-12-20" "1950-01-19" "1950-02-19" "1950-03-22" "1950-04-19"
#>  [16] "1950-05-20" "1950-06-19" "1950-07-20" "1950-08-19" "1950-09-19"
#>  [21] "1950-10-20" "1950-11-19" "1950-12-20" "1951-01-19" "1951-02-19"
#>  [26] "1951-03-22" "1951-04-19" "1951-05-20" "1951-06-19" "1951-07-20"
#>  [31] "1951-08-19" "1951-09-19" "1951-10-20" "1951-11-19" "1951-12-20"
#>  [36] "1952-01-19" "1952-02-19" "1952-03-21" "1952-04-19" "1952-05-20"
#>  [41] "1952-06-19" "1952-07-20" "1952-08-19" "1952-09-19" "1952-10-20"
#>  [46] "1952-11-19" "1952-12-20" "1953-01-19" "1953-02-19" "1953-03-22"
#>  [51] "1953-04-19" "1953-05-20" "1953-06-19" "1953-07-20" "1953-08-19"
#>  [56] "1953-09-19" "1953-10-20" "1953-11-19" "1953-12-20" "1954-01-19"
#>  [61] "1954-02-19" "1954-03-22" "1954-04-19" "1954-05-20" "1954-06-19"
#>  [66] "1954-07-20" "1954-08-19" "1954-09-19" "1954-10-20" "1954-11-19"
#>  [71] "1954-12-20" "1955-01-19" "1955-02-19" "1955-03-22" "1955-04-19"
#>  [76] "1955-05-20" "1955-06-19" "1955-07-20" "1955-08-19" "1955-09-19"
#>  [81] "1955-10-20" "1955-11-19" "1955-12-20" "1956-01-19" "1956-02-19"
#>  [86] "1956-03-21" "1956-04-19" "1956-05-20" "1956-06-19" "1956-07-20"
#>  [91] "1956-08-19" "1956-09-19" "1956-10-20" "1956-11-19" "1956-12-20"
#>  [96] "1957-01-19" "1957-02-19" "1957-03-22" "1957-04-19" "1957-05-20"
#> [101] "1957-06-19" "1957-07-20" "1957-08-19" "1957-09-19" "1957-10-20"
#> [106] "1957-11-19" "1957-12-20" "1958-01-19" "1958-02-19" "1958-03-22"
#> [111] "1958-04-19" "1958-05-20" "1958-06-19" "1958-07-20" "1958-08-19"
#> [116] "1958-09-19" "1958-10-20" "1958-11-19" "1958-12-20" "1959-01-19"
#> [121] "1959-02-19" "1959-03-22" "1959-04-19" "1959-05-20" "1959-06-19"
#> [126] "1959-07-20" "1959-08-19" "1959-09-19" "1959-10-20" "1959-11-19"
#> [131] "1959-12-20" "1960-01-19" "1960-02-19" "1960-03-21" "1960-04-19"
#> [136] "1960-05-20" "1960-06-19" "1960-07-20" "1960-08-19" "1960-09-19"
#> [141] "1960-10-20" "1960-11-19" "1960-12-20" "1961-01-19"

date_shift(ts_df(AirPassengers)$time, "-1 month")

#>   [1] "1948-12-01" "1949-01-01" "1949-02-01" "1949-03-01" "1949-04-01"
#>   [6] "1949-05-01" "1949-06-01" "1949-07-01" "1949-08-01" "1949-09-01"
#>  [11] "1949-10-01" "1949-11-01" "1949-12-01" "1950-01-01" "1950-02-01"
#>  [16] "1950-03-01" "1950-04-01" "1950-05-01" "1950-06-01" "1950-07-01"
#>  [21] "1950-08-01" "1950-09-01" "1950-10-01" "1950-11-01" "1950-12-01"
#>  [26] "1951-01-01" "1951-02-01" "1951-03-01" "1951-04-01" "1951-05-01"
#>  [31] "1951-06-01" "1951-07-01" "1951-08-01" "1951-09-01" "1951-10-01"
#>  [36] "1951-11-01" "1951-12-01" "1952-01-01" "1952-02-01" "1952-03-01"
#>  [41] "1952-04-01" "1952-05-01" "1952-06-01" "1952-07-01" "1952-08-01"
#>  [46] "1952-09-01" "1952-10-01" "1952-11-01" "1952-12-01" "1953-01-01"
#>  [51] "1953-02-01" "1953-03-01" "1953-04-01" "1953-05-01" "1953-06-01"
#>  [56] "1953-07-01" "1953-08-01" "1953-09-01" "1953-10-01" "1953-11-01"
#>  [61] "1953-12-01" "1954-01-01" "1954-02-01" "1954-03-01" "1954-04-01"
#>  [66] "1954-05-01" "1954-06-01" "1954-07-01" "1954-08-01" "1954-09-01"
#>  [71] "1954-10-01" "1954-11-01" "1954-12-01" "1955-01-01" "1955-02-01"
#>  [76] "1955-03-01" "1955-04-01" "1955-05-01" "1955-06-01" "1955-07-01"
#>  [81] "1955-08-01" "1955-09-01" "1955-10-01" "1955-11-01" "1955-12-01"
#>  [86] "1956-01-01" "1956-02-01" "1956-03-01" "1956-04-01" "1956-05-01"
#>  [91] "1956-06-01" "1956-07-01" "1956-08-01" "1956-09-01" "1956-10-01"
#>  [96] "1956-11-01" "1956-12-01" "1957-01-01" "1957-02-01" "1957-03-01"
#> [101] "1957-04-01" "1957-05-01" "1957-06-01" "1957-07-01" "1957-08-01"
#> [106] "1957-09-01" "1957-10-01" "1957-11-01" "1957-12-01" "1958-01-01"
#> [111] "1958-02-01" "1958-03-01" "1958-04-01" "1958-05-01" "1958-06-01"
#> [116] "1958-07-01" "1958-08-01" "1958-09-01" "1958-10-01" "1958-11-01"
#> [121] "1958-12-01" "1959-01-01" "1959-02-01" "1959-03-01" "1959-04-01"
#> [126] "1959-05-01" "1959-06-01" "1959-07-01" "1959-08-01" "1959-09-01"
#> [131] "1959-10-01" "1959-11-01" "1959-12-01" "1960-01-01" "1960-02-01"
#> [136] "1960-03-01" "1960-04-01" "1960-05-01" "1960-06-01" "1960-07-01"
#> [141] "1960-08-01" "1960-09-01" "1960-10-01" "1960-11-01"