Indices from Levels or Percentage Rates

ts_index returns an index series, with value of 1 at base date. ts_compound builds an index from percentage change rates, starting with 1 and compounding the rates.

ts_compound(x, denominator = 100)

ts_index(x, base)

Arguments

x
ts-boxable time series, an object of class ts, xts, data.frame, data.table, or tibble.

denominator
numeric, set equal to one if percentage change rate is given a decimal fraction

base
base date, character string, Date or POSIXct, at which the

Value

a ts-boxable time series, with the same class as the input.

Examples

ts_compound(ts_pc(ts_c(fdeaths, mdeaths)))

#>            fdeaths   mdeaths
#> Feb 1974 0.7647059 0.8730084
#> Mar 1974 0.9178690 0.8795689
#> Apr 1974 0.7513873 0.8795689
#> May 1974 0.5793563 0.6991565
#> Jun 1974 0.4506104 0.5852858
#> Jul 1974 0.4894562 0.5998126
#> Aug 1974 0.4361820 0.5299906
#> Sep 1974 0.4295228 0.5665417
#> Oct 1974 0.6459489 0.6991565
#> Nov 1974 0.6415094 0.7596064
#> Dec 1974 0.7391787 0.8650422
#> Jan 1975 0.9211987 0.9854733
#> Feb 1975 0.8346282 1.0014058
#> Mar 1975 0.8712542 1.0089035
#> Apr 1975 0.7369589 0.8589503
#> May 1975 0.5183130 0.6574508
#> Jun 1975 0.4861265 0.6035614
#> Jul 1975 0.4672586 0.5557638
#> Aug 1975 0.4572697 0.5309278
#> Sep 1975 0.3806881 0.4934396
#> Oct 1975 0.4883463 0.6312090
#> Nov 1975 0.5893452 0.7239925
#> Dec 1975 0.8557159 0.9681350
#> Jan 1976 0.8512764 0.9465792
#> Feb 1976 1.2663707 1.2886598
#> Mar 1976 0.9944506 1.0698219
#> Apr 1976 0.5904550 0.6930647
#> May 1976 0.4961154 0.5571696
#> Jun 1976 0.4661487 0.5435801
#> Jul 1976 0.4173141 0.5215558
#> Aug 1976 0.3662597 0.4545455
#> Sep 1976 0.3962264 0.4681350
#> Oct 1976 0.4938957 0.5660731
#> Nov 1976 0.6059933 0.6874414
#> Dec 1976 0.8479467 0.9648547
#> Jan 1977 0.9567148 1.0496720
#> Feb 1977 0.7325194 0.7656982
#> Mar 1977 0.7358491 0.8069353
#> Apr 1977 0.7136515 0.8439550
#> May 1977 0.5571587 0.5838800
#> Jun 1977 0.4350721 0.5445173
#> Jul 1977 0.4561598 0.5093721
#> Aug 1977 0.3862375 0.4746954
#> Sep 1977 0.4295228 0.4493908
#> Oct 1977 0.4273030 0.5524836
#> Nov 1977 0.4561598 0.5759138
#> Dec 1977 0.7081021 0.7755389
#> Jan 1978 0.8834628 0.9461106
#> Feb 1978 0.9467259 1.0702905
#> Mar 1978 0.8179800 0.9100281
#> Apr 1978 0.6059933 0.6668229
#> May 1978 0.5882353 0.6279288
#> Jun 1978 0.4950055 0.5562324
#> Jul 1978 0.4783574 0.5145267
#> Aug 1978 0.4017758 0.4704780
#> Sep 1978 0.4295228 0.4545455
#> Oct 1978 0.4772475 0.5342081
#> Nov 1978 0.4716981 0.5201500
#> Dec 1978 0.7536071 0.8491097
#> Jan 1979 0.9112098 1.0604499
#> Feb 1979 0.8712542 0.8528585
#> Mar 1979 0.8068812 0.8650422
#> Apr 1979 0.6792453 0.7174321
#> May 1979 0.5305216 0.5693533
#> Jun 1979 0.4761376 0.5037488
#> Jul 1979 0.4495006 0.4948454
#> Aug 1979 0.4206437 0.4568885
#> Sep 1979 0.4361820 0.4404873
#> Oct 1979 0.4561598 0.5065604
#> Nov 1979 0.5405105 0.6063730
#> Dec 1979 0.6370699 0.6283974

ts_index(ts_df(ts_c(fdeaths, mdeaths)), "1977-01-01")

#>          id       time     value
#> 1   fdeaths 1974-01-01 1.0452436
#> 2   fdeaths 1974-02-01 0.7993039
#> 3   fdeaths 1974-03-01 0.9593968
#> 4   fdeaths 1974-04-01 0.7853828
#> 5   fdeaths 1974-05-01 0.6055684
#> 6   fdeaths 1974-06-01 0.4709977
#> 7   fdeaths 1974-07-01 0.5116009
#> 8   fdeaths 1974-08-01 0.4559165
#> 9   fdeaths 1974-09-01 0.4489559
#> 10  fdeaths 1974-10-01 0.6751740
#> 11  fdeaths 1974-11-01 0.6705336
#> 12  fdeaths 1974-12-01 0.7726218
#> 13  fdeaths 1975-01-01 0.9628770
#> 14  fdeaths 1975-02-01 0.8723898
#> 15  fdeaths 1975-03-01 0.9106729
#> 16  fdeaths 1975-04-01 0.7703016
#> 17  fdeaths 1975-05-01 0.5417633
#> 18  fdeaths 1975-06-01 0.5081206
#> 19  fdeaths 1975-07-01 0.4883991
#> 20  fdeaths 1975-08-01 0.4779582
#> 21  fdeaths 1975-09-01 0.3979118
#> 22  fdeaths 1975-10-01 0.5104408
#> 23  fdeaths 1975-11-01 0.6160093
#> 24  fdeaths 1975-12-01 0.8944316
#> 25  fdeaths 1976-01-01 0.8897912
#> 26  fdeaths 1976-02-01 1.3236659
#> 27  fdeaths 1976-03-01 1.0394432
#> 28  fdeaths 1976-04-01 0.6171694
#> 29  fdeaths 1976-05-01 0.5185615
#> 30  fdeaths 1976-06-01 0.4872390
#> 31  fdeaths 1976-07-01 0.4361949
#> 32  fdeaths 1976-08-01 0.3828306
#> 33  fdeaths 1976-09-01 0.4141531
#> 34  fdeaths 1976-10-01 0.5162413
#> 35  fdeaths 1976-11-01 0.6334107
#> 36  fdeaths 1976-12-01 0.8863109
#> 37  fdeaths 1977-01-01 1.0000000
#> 38  fdeaths 1977-02-01 0.7656613
#> 39  fdeaths 1977-03-01 0.7691415
#> 40  fdeaths 1977-04-01 0.7459397
#> 41  fdeaths 1977-05-01 0.5823666
#> 42  fdeaths 1977-06-01 0.4547564
#> 43  fdeaths 1977-07-01 0.4767981
#> 44  fdeaths 1977-08-01 0.4037123
#> 45  fdeaths 1977-09-01 0.4489559
#> 46  fdeaths 1977-10-01 0.4466357
#> 47  fdeaths 1977-11-01 0.4767981
#> 48  fdeaths 1977-12-01 0.7401392
#> 49  fdeaths 1978-01-01 0.9234339
#> 50  fdeaths 1978-02-01 0.9895592
#> 51  fdeaths 1978-03-01 0.8549884
#> 52  fdeaths 1978-04-01 0.6334107
#> 53  fdeaths 1978-05-01 0.6148492
#> 54  fdeaths 1978-06-01 0.5174014
#> 55  fdeaths 1978-07-01 0.5000000
#> 56  fdeaths 1978-08-01 0.4199536
#> 57  fdeaths 1978-09-01 0.4489559
#> 58  fdeaths 1978-10-01 0.4988399
#> 59  fdeaths 1978-11-01 0.4930394
#> 60  fdeaths 1978-12-01 0.7877030
#> 61  fdeaths 1979-01-01 0.9524362
#> 62  fdeaths 1979-02-01 0.9106729
#> 63  fdeaths 1979-03-01 0.8433875
#> 64  fdeaths 1979-04-01 0.7099768
#> 65  fdeaths 1979-05-01 0.5545244
#> 66  fdeaths 1979-06-01 0.4976798
#> 67  fdeaths 1979-07-01 0.4698376
#> 68  fdeaths 1979-08-01 0.4396752
#> 69  fdeaths 1979-09-01 0.4559165
#> 70  fdeaths 1979-10-01 0.4767981
#> 71  fdeaths 1979-11-01 0.5649652
#> 72  fdeaths 1979-12-01 0.6658933
#> 73  mdeaths 1974-01-01 0.9526786
#> 74  mdeaths 1974-02-01 0.8316964
#> 75  mdeaths 1974-03-01 0.8379464
#> 76  mdeaths 1974-04-01 0.8379464
#> 77  mdeaths 1974-05-01 0.6660714
#> 78  mdeaths 1974-06-01 0.5575893
#> 79  mdeaths 1974-07-01 0.5714286
#> 80  mdeaths 1974-08-01 0.5049107
#> 81  mdeaths 1974-09-01 0.5397321
#> 82  mdeaths 1974-10-01 0.6660714
#> 83  mdeaths 1974-11-01 0.7236607
#> 84  mdeaths 1974-12-01 0.8241071
#> 85  mdeaths 1975-01-01 0.9388393
#> 86  mdeaths 1975-02-01 0.9540179
#> 87  mdeaths 1975-03-01 0.9611607
#> 88  mdeaths 1975-04-01 0.8183036
#> 89  mdeaths 1975-05-01 0.6263393
#> 90  mdeaths 1975-06-01 0.5750000
#> 91  mdeaths 1975-07-01 0.5294643
#> 92  mdeaths 1975-08-01 0.5058036
#> 93  mdeaths 1975-09-01 0.4700893
#> 94  mdeaths 1975-10-01 0.6013393
#> 95  mdeaths 1975-11-01 0.6897321
#> 96  mdeaths 1975-12-01 0.9223214
#> 97  mdeaths 1976-01-01 0.9017857
#> 98  mdeaths 1976-02-01 1.2276786
#> 99  mdeaths 1976-03-01 1.0191964
#> 100 mdeaths 1976-04-01 0.6602679
#> 101 mdeaths 1976-05-01 0.5308036
#> 102 mdeaths 1976-06-01 0.5178571
#> 103 mdeaths 1976-07-01 0.4968750
#> 104 mdeaths 1976-08-01 0.4330357
#> 105 mdeaths 1976-09-01 0.4459821
#> 106 mdeaths 1976-10-01 0.5392857
#> 107 mdeaths 1976-11-01 0.6549107
#> 108 mdeaths 1976-12-01 0.9191964
#> 109 mdeaths 1977-01-01 1.0000000
#> 110 mdeaths 1977-02-01 0.7294643
#> 111 mdeaths 1977-03-01 0.7687500
#> 112 mdeaths 1977-04-01 0.8040179
#> 113 mdeaths 1977-05-01 0.5562500
#> 114 mdeaths 1977-06-01 0.5187500
#> 115 mdeaths 1977-07-01 0.4852679
#> 116 mdeaths 1977-08-01 0.4522321
#> 117 mdeaths 1977-09-01 0.4281250
#> 118 mdeaths 1977-10-01 0.5263393
#> 119 mdeaths 1977-11-01 0.5486607
#> 120 mdeaths 1977-12-01 0.7388393
#> 121 mdeaths 1978-01-01 0.9013393
#> 122 mdeaths 1978-02-01 1.0196429
#> 123 mdeaths 1978-03-01 0.8669643
#> 124 mdeaths 1978-04-01 0.6352679
#> 125 mdeaths 1978-05-01 0.5982143
#> 126 mdeaths 1978-06-01 0.5299107
#> 127 mdeaths 1978-07-01 0.4901786
#> 128 mdeaths 1978-08-01 0.4482143
#> 129 mdeaths 1978-09-01 0.4330357
#> 130 mdeaths 1978-10-01 0.5089286
#> 131 mdeaths 1978-11-01 0.4955357
#> 132 mdeaths 1978-12-01 0.8089286
#> 133 mdeaths 1979-01-01 1.0102679
#> 134 mdeaths 1979-02-01 0.8125000
#> 135 mdeaths 1979-03-01 0.8241071
#> 136 mdeaths 1979-04-01 0.6834821
#> 137 mdeaths 1979-05-01 0.5424107
#> 138 mdeaths 1979-06-01 0.4799107
#> 139 mdeaths 1979-07-01 0.4714286
#> 140 mdeaths 1979-08-01 0.4352679
#> 141 mdeaths 1979-09-01 0.4196429
#> 142 mdeaths 1979-10-01 0.4825893
#> 143 mdeaths 1979-11-01 0.5776786
#> 144 mdeaths 1979-12-01 0.5986607
# NOT RUN {
ts_plot(
  `My Expert Knowledge` = ts_chain(mdeaths, ts_compound(ts_bind(ts_pc(mdeaths), 15, 23, 33))),
  `So Far` = mdeaths,
  title = "A Very Manual Forecast"
)
# }