Principal Components, Dygraphs, Forecasts, Seasonal Adjustment

Example Functions, Generated by ts_. ts_prcomp calculates the principal components of multiple time series, ts_dygraphs generates an interactive graphical visualization, ts_forecast return an univariate forecast, ts_seas the seasonally adjusted series.

ts_prcomp(x, ...)

ts_dygraphs(x, ...)

ts_forecast(x, ...)

ts_seas(x, ...)

Arguments

x
ts-boxable time series, an object of class ts, xts, data.frame, data.table, or tibble.

...
further arguments, passed to the underlying function. For help, consider these functions, e.g., stats::prcomp.

Value

Usually, a ts-boxable time series, with the same class as the input. ts_dygraphs draws a plot.

Details

With the exception of ts_prcomp, these functions depend on external packages.

Examples

ts_prcomp(ts_c(mdeaths, fdeaths))

#>                  PC1          PC2
#> Jan 1974 -2.38059153  0.297375755
#> Feb 1974 -1.10407959 -0.094336308
#> Mar 1974 -1.66989390  0.425768781
#> Apr 1974 -1.07972032 -0.164404804
#> May 1974  0.15862748 -0.145749085
#> Jun 1974  1.01171910 -0.205459272
#> Jul 1974  0.82340530 -0.118358737
#> Aug 1974  1.25549919 -0.063975943
#> Sep 1974  1.15177331 -0.214915709
#> Oct 1974 -0.07744196  0.090320349
#> Nov 1974 -0.27229290 -0.136006514
#> Dec 1974 -0.98583429 -0.157077565
#> Jan 1975 -2.05063607  0.068633560
#> Feb 1975 -1.79924985 -0.293760756
#> Mar 1975 -1.95520760 -0.190042121
#> Apr 1975 -0.95674317 -0.143724409
#> May 1975  0.52031447 -0.216856050
#> Jun 1975  0.82214899 -0.143221986
#> Jul 1975  1.05554748 -0.043596172
#> Aug 1975  1.17747892  0.007514434
#> Sep 1975  1.57955654 -0.133367648
#> Oct 1975  0.71796416 -0.231668858
#> Nov 1975  0.03669604 -0.196859691
#> Dec 1975 -1.75809966 -0.103099916
#> Jan 1976 -1.66726798 -0.043744162
#> Feb 1976 -4.33047208  0.236050681
#> Mar 1976 -2.60415742  0.034464960
#> Apr 1976  0.14050471 -0.085182043
#> May 1976  0.94835331  0.053803167
#> Jun 1976  1.10192625 -0.005086388
#> Jul 1976  1.35177002 -0.101477785
#> Aug 1976  1.76620010 -0.049020841
#> Sep 1976  1.61262716  0.009868713
#> Oct 1976  0.92520532  0.014917216
#> Nov 1976  0.10501151 -0.010509510
#> Dec 1976 -1.71913092 -0.119214046
#> Jan 1977 -2.40018844 -0.029111419
#> Feb 1977 -0.61614326  0.165399575
#> Mar 1977 -0.77160427  0.033545503
#> Apr 1977 -0.82187976 -0.174109604
#> May 1977  0.63890542  0.177149238
#> Jun 1977  1.20882704 -0.118517068
#> Jul 1977  1.25650713  0.078673660
#> Aug 1977  1.62518297 -0.048396311
#> Sep 1977  1.55989133  0.193202314
#> Oct 1977  1.20861645 -0.173810528
#> Nov 1977  1.02469609 -0.153137378
#> Dec 1977 -0.56386638  0.044558869
#> Jan 1978 -1.77973573  0.071988536
#> Feb 1978 -2.43660680 -0.136350606
#> Mar 1978 -1.42190044 -0.034446056
#> Apr 1978  0.17684028  0.061319263
#> May 1978  0.37528731  0.133862597
#> Jun 1978  0.95555275  0.053133620
#> Jul 1978  1.15986012  0.139406278
#> Aug 1978  1.58479235  0.021378806
#> Sep 1978  1.54193414  0.175245121
#> Oct 1978  1.09523079  0.066907960
#> Nov 1978  1.16387740  0.096209669
#> Dec 1978 -0.98147861 -0.050425137
#> Jan 1979 -2.27642119 -0.227972391
#> Feb 1979 -1.41159439  0.353571087
#> Mar 1979 -1.22583821  0.082926359
#> Apr 1979 -0.25914309  0.144688654
#> May 1979  0.78393983  0.133328100
#> Jun 1979  1.20527596  0.169084155
#> Jul 1979  1.33072071  0.105673351
#> Aug 1979  1.56524770  0.135606836
#> Sep 1979  1.56730136  0.247826227
#> Oct 1979  1.26630196  0.088468492
#> Nov 1979  0.61956412  0.039773219
#> Dec 1979  0.20053725  0.305347710

ts_dygraphs(ts_c(mdeaths, EuStockMarkets))
ts_forecast(ts_c(mdeaths, fdeaths))

#>            mdeaths  fdeaths
#> Jan 1980 1893.1855 798.7897
#> Feb 1980 1886.6774 818.3414
#> Mar 1980 1750.3776 748.4718
#> Apr 1980 1494.0642 593.0345
#> May 1980 1171.2965 478.9029
#> Jun 1980 1068.0515 406.1266
#> Jul 1980 1008.9518 401.9049
#> Aug 1980  924.6615 359.6939
#> Sep 1980  914.5511 366.6550
#> Oct 1980 1098.6821 439.1029
#> Nov 1980 1235.0739 487.9075
#> Dec 1980 1636.0542 673.0117
#> Jan 1981 1893.1857 798.7897
#> Feb 1981 1886.6777 818.3414
#> Mar 1981 1750.3779 748.4718
#> Apr 1981 1494.0644 593.0345
#> May 1981 1171.2967 478.9029
#> Jun 1981 1068.0517 406.1266
#> Jul 1981 1008.9519 401.9049
#> Aug 1981  924.6616 359.6939
#> Sep 1981  914.5513 366.6550
#> Oct 1981 1098.6823 439.1030
#> Nov 1981 1235.0741 487.9076
#> Dec 1981 1636.0544 673.0117

ts_seas(ts_c(mdeaths, fdeaths))

#>           mdeaths  fdeaths
#> Jan 1974 1520.435 605.9811
#> Feb 1974 1344.908 538.6371
#> Mar 1974 1423.483 589.4354
#> Apr 1974 1722.832 606.7556
#> May 1974 1715.809 599.5486
#> Jun 1974 1575.049 547.2017
#> Jul 1974 1682.336 592.4375
#> Aug 1974 1627.528 583.6646
#> Sep 1974 1761.674 585.0864
#> Oct 1974 1781.126 682.7610
#> Nov 1974 1752.479 617.8549
#> Dec 1974 1493.939 534.4683
#> Jan 1975 1498.332 561.6058
#> Feb 1975 1542.352 578.3942
#> Mar 1975 1632.551 559.8050
#> Apr 1975 1682.542 599.4746
#> May 1975 1613.539 538.0181
#> Jun 1975 1624.220 582.5884
#> Jul 1975 1558.779 566.5253
#> Aug 1975 1630.400 611.9697
#> Sep 1975 1534.395 517.3097
#> Oct 1975 1608.123 530.8710
#> Nov 1975 1670.322 573.5895
#> Dec 1975 1672.360 608.8541
#> Jan 1976 1439.167 518.7201
#> Feb 1976 1984.310 862.1765
#> Mar 1976 1730.865 637.6972
#> Apr 1976 1357.680 486.6755
#> May 1976 1367.494 510.4945
#> Jun 1976 1462.797 552.8565
#> Jul 1976 1462.821 506.1075
#> Aug 1976 1395.835 496.4607
#> Sep 1976 1455.737 529.3762
#> Oct 1976 1442.263 549.1862
#> Nov 1976 1586.013 601.2601
#> Dec 1976 1667.073 604.1325
#> Jan 1977 1595.867 576.8622
#> Feb 1977 1178.784 490.6124
#> Mar 1977 1305.357 482.5705
#> Apr 1977 1653.348 585.7041
#> May 1977 1433.106 562.1201
#> Jun 1977 1465.307 512.3556
#> Jul 1977 1428.632 547.2420
#> Aug 1977 1457.699 522.9962
#> Sep 1977 1397.472 561.7619
#> Oct 1977 1407.721 485.8909
#> Nov 1977 1328.726 469.0500
#> Dec 1977 1340.299 515.5852
#> Jan 1978 1438.388 532.7679
#> Feb 1978 1647.319 615.6214
#> Mar 1978 1471.908 540.0766
#> Apr 1978 1306.416 501.3649
#> May 1978 1541.284 587.1514
#> Jun 1978 1496.820 575.3995
#> Jul 1978 1443.073 570.0180
#> Aug 1978 1444.736 540.7830
#> Sep 1978 1413.522 555.1463
#> Oct 1978 1361.232 546.4384
#> Nov 1978 1200.082 491.7076
#> Dec 1978 1467.791 558.8122
#> Jan 1979 1612.192 549.7471
#> Feb 1979 1312.370 562.4985
#> Mar 1979 1398.941 535.0491
#> Apr 1979 1405.642 559.8752
#> May 1979 1397.571 531.1366
#> Jun 1979 1355.583 551.0426
#> Jul 1979 1387.858 535.9516
#> Aug 1979 1402.995 561.6277
#> Sep 1979 1369.826 559.1184
#> Oct 1979 1290.848 524.3141
#> Nov 1979 1399.010 559.2716
#> Dec 1979 1086.521 481.2997