Lags, Differences, Percentage Change Rates

Utlility functions to work with time series. ts_pcy and ts_diffy calulate the percentage change and the difference compared to the same period of the previous year, rather than the previous period. ts_index returns an index series, with value of 1 at base date.

ts_lag(x, lag = 1, fill = NA)

ts_pc(x)

ts_diff(x)

ts_pcy(x, ...)

ts_diffy(x, ...)

Arguments

x
ts-boxable time series, an object of class ts, xts, data.frame, data.table, or tibble.

lag
integer, the number of lags. This is defined as in dplyr, opposite to R base.

fill
how to fill missing values

...
additional arguments, passed to subfunctions (as some of these functions are generated by ts_)

Value

a ts-boxable time series, with the same class as the input.

Examples

ts_lag(ts_c(fdeaths, mdeaths))

#>          fdeaths mdeaths
#> Feb 1974     901    2134
#> Mar 1974     689    1863
#> Apr 1974     827    1877
#> May 1974     677    1877
#> Jun 1974     522    1492
#> Jul 1974     406    1249
#> Aug 1974     441    1280
#> Sep 1974     393    1131
#> Oct 1974     387    1209
#> Nov 1974     582    1492
#> Dec 1974     578    1621
#> Jan 1975     666    1846
#> Feb 1975     830    2103
#> Mar 1975     752    2137
#> Apr 1975     785    2153
#> May 1975     664    1833
#> Jun 1975     467    1403
#> Jul 1975     438    1288
#> Aug 1975     421    1186
#> Sep 1975     412    1133
#> Oct 1975     343    1053
#> Nov 1975     440    1347
#> Dec 1975     531    1545
#> Jan 1976     771    2066
#> Feb 1976     767    2020
#> Mar 1976    1141    2750
#> Apr 1976     896    2283
#> May 1976     532    1479
#> Jun 1976     447    1189
#> Jul 1976     420    1160
#> Aug 1976     376    1113
#> Sep 1976     330     970
#> Oct 1976     357     999
#> Nov 1976     445    1208
#> Dec 1976     546    1467
#> Jan 1977     764    2059
#> Feb 1977     862    2240
#> Mar 1977     660    1634
#> Apr 1977     663    1722
#> May 1977     643    1801
#> Jun 1977     502    1246
#> Jul 1977     392    1162
#> Aug 1977     411    1087
#> Sep 1977     348    1013
#> Oct 1977     387     959
#> Nov 1977     385    1179
#> Dec 1977     411    1229
#> Jan 1978     638    1655
#> Feb 1978     796    2019
#> Mar 1978     853    2284
#> Apr 1978     737    1942
#> May 1978     546    1423
#> Jun 1978     530    1340
#> Jul 1978     446    1187
#> Aug 1978     431    1098
#> Sep 1978     362    1004
#> Oct 1978     387     970
#> Nov 1978     430    1140
#> Dec 1978     425    1110
#> Jan 1979     679    1812
#> Feb 1979     821    2263
#> Mar 1979     785    1820
#> Apr 1979     727    1846
#> May 1979     612    1531
#> Jun 1979     478    1215
#> Jul 1979     429    1075
#> Aug 1979     405    1056
#> Sep 1979     379     975
#> Oct 1979     393     940
#> Nov 1979     411    1081
#> Dec 1979     487    1294

ts_diff(ts_c(fdeaths, mdeaths))

#>               fdeaths      mdeaths
#> Feb 1974 -0.235294118 -0.126991565
#> Mar 1974  0.200290276  0.007514761
#> Apr 1974 -0.181378476  0.000000000
#> May 1974 -0.228951256 -0.205114544
#> Jun 1974 -0.222222222 -0.162868633
#> Jul 1974  0.086206897  0.024819856
#> Aug 1974 -0.108843537 -0.116406250
#> Sep 1974 -0.015267176  0.068965517
#> Oct 1974  0.503875969  0.234077750
#> Nov 1974 -0.006872852  0.086461126
#> Dec 1974  0.152249135  0.138803208
#> Jan 1975  0.246246246  0.139219935
#> Feb 1975 -0.093975904  0.016167380
#> Mar 1975  0.043882979  0.007487131
#> Apr 1975 -0.154140127 -0.148629819
#> May 1975 -0.296686747 -0.234588107
#> Jun 1975 -0.062098501 -0.081967213
#> Jul 1975 -0.038812785 -0.079192547
#> Aug 1975 -0.021377672 -0.044688027
#> Sep 1975 -0.167475728 -0.070609003
#> Oct 1975  0.282798834  0.279202279
#> Nov 1975  0.206818182  0.146993318
#> Dec 1975  0.451977401  0.337216828
#> Jan 1976 -0.005188067 -0.022265247
#> Feb 1976  0.487614081  0.361386139
#> Mar 1976 -0.214723926 -0.169818182
#> Apr 1976 -0.406250000 -0.352168200
#> May 1976 -0.159774436 -0.196078431
#> Jun 1976 -0.060402685 -0.024390244
#> Jul 1976 -0.104761905 -0.040517241
#> Aug 1976 -0.122340426 -0.128481581
#> Sep 1976  0.081818182  0.029896907
#> Oct 1976  0.246498599  0.209209209
#> Nov 1976  0.226966292  0.214403974
#> Dec 1976  0.399267399  0.403544649
#> Jan 1977  0.128272251  0.087906751
#> Feb 1977 -0.234338747 -0.270535714
#> Mar 1977  0.004545455  0.053855569
#> Apr 1977 -0.030165913  0.045876887
#> May 1977 -0.219284603 -0.308162132
#> Jun 1977 -0.219123506 -0.067415730
#> Jul 1977  0.048469388 -0.064543890
#> Aug 1977 -0.153284672 -0.068077277
#> Sep 1977  0.112068966 -0.053307009
#> Oct 1977 -0.005167959  0.229405631
#> Nov 1977  0.067532468  0.042408821
#> Dec 1977  0.552311436  0.346623271
#> Jan 1978  0.247648903  0.219939577
#> Feb 1978  0.071608040  0.131253096
#> Mar 1978 -0.135990621 -0.149737303
#> Apr 1978 -0.259158752 -0.267250257
#> May 1978 -0.029304029 -0.058327477
#> Jun 1978 -0.158490566 -0.114179104
#> Jul 1978 -0.033632287 -0.074978939
#> Aug 1978 -0.160092807 -0.085610200
#> Sep 1978  0.069060773 -0.033864542
#> Oct 1978  0.111111111  0.175257732
#> Nov 1978 -0.011627907 -0.026315789
#> Dec 1978  0.597647059  0.632432432
#> Jan 1979  0.209131075  0.248896247
#> Feb 1979 -0.043848965 -0.195757844
#> Mar 1979 -0.073885350  0.014285714
#> Apr 1979 -0.158184319 -0.170639220
#> May 1979 -0.218954248 -0.206401045
#> Jun 1979 -0.102510460 -0.115226337
#> Jul 1979 -0.055944056 -0.017674419
#> Aug 1979 -0.064197531 -0.076704545
#> Sep 1979  0.036939314 -0.035897436
#> Oct 1979  0.045801527  0.150000000
#> Nov 1979  0.184914842  0.197039778
#> Dec 1979  0.178644764  0.036321484

ts_pc(ts_c(fdeaths, mdeaths))

#>              fdeaths     mdeaths
#> Feb 1974 -23.5294118 -12.6991565
#> Mar 1974  20.0290276   0.7514761
#> Apr 1974 -18.1378476   0.0000000
#> May 1974 -22.8951256 -20.5114544
#> Jun 1974 -22.2222222 -16.2868633
#> Jul 1974   8.6206897   2.4819856
#> Aug 1974 -10.8843537 -11.6406250
#> Sep 1974  -1.5267176   6.8965517
#> Oct 1974  50.3875969  23.4077750
#> Nov 1974  -0.6872852   8.6461126
#> Dec 1974  15.2249135  13.8803208
#> Jan 1975  24.6246246  13.9219935
#> Feb 1975  -9.3975904   1.6167380
#> Mar 1975   4.3882979   0.7487131
#> Apr 1975 -15.4140127 -14.8629819
#> May 1975 -29.6686747 -23.4588107
#> Jun 1975  -6.2098501  -8.1967213
#> Jul 1975  -3.8812785  -7.9192547
#> Aug 1975  -2.1377672  -4.4688027
#> Sep 1975 -16.7475728  -7.0609003
#> Oct 1975  28.2798834  27.9202279
#> Nov 1975  20.6818182  14.6993318
#> Dec 1975  45.1977401  33.7216828
#> Jan 1976  -0.5188067  -2.2265247
#> Feb 1976  48.7614081  36.1386139
#> Mar 1976 -21.4723926 -16.9818182
#> Apr 1976 -40.6250000 -35.2168200
#> May 1976 -15.9774436 -19.6078431
#> Jun 1976  -6.0402685  -2.4390244
#> Jul 1976 -10.4761905  -4.0517241
#> Aug 1976 -12.2340426 -12.8481581
#> Sep 1976   8.1818182   2.9896907
#> Oct 1976  24.6498599  20.9209209
#> Nov 1976  22.6966292  21.4403974
#> Dec 1976  39.9267399  40.3544649
#> Jan 1977  12.8272251   8.7906751
#> Feb 1977 -23.4338747 -27.0535714
#> Mar 1977   0.4545455   5.3855569
#> Apr 1977  -3.0165913   4.5876887
#> May 1977 -21.9284603 -30.8162132
#> Jun 1977 -21.9123506  -6.7415730
#> Jul 1977   4.8469388  -6.4543890
#> Aug 1977 -15.3284672  -6.8077277
#> Sep 1977  11.2068966  -5.3307009
#> Oct 1977  -0.5167959  22.9405631
#> Nov 1977   6.7532468   4.2408821
#> Dec 1977  55.2311436  34.6623271
#> Jan 1978  24.7648903  21.9939577
#> Feb 1978   7.1608040  13.1253096
#> Mar 1978 -13.5990621 -14.9737303
#> Apr 1978 -25.9158752 -26.7250257
#> May 1978  -2.9304029  -5.8327477
#> Jun 1978 -15.8490566 -11.4179104
#> Jul 1978  -3.3632287  -7.4978939
#> Aug 1978 -16.0092807  -8.5610200
#> Sep 1978   6.9060773  -3.3864542
#> Oct 1978  11.1111111  17.5257732
#> Nov 1978  -1.1627907  -2.6315789
#> Dec 1978  59.7647059  63.2432432
#> Jan 1979  20.9131075  24.8896247
#> Feb 1979  -4.3848965 -19.5757844
#> Mar 1979  -7.3885350   1.4285714
#> Apr 1979 -15.8184319 -17.0639220
#> May 1979 -21.8954248 -20.6401045
#> Jun 1979 -10.2510460 -11.5226337
#> Jul 1979  -5.5944056  -1.7674419
#> Aug 1979  -6.4197531  -7.6704545
#> Sep 1979   3.6939314  -3.5897436
#> Oct 1979   4.5801527  15.0000000
#> Nov 1979  18.4914842  19.7039778
#> Dec 1979  17.8644764   3.6321484