Shift Time Stamp

Shift time stamps in ts-boxable time series. Unlike ts_lag this does not ensure regularity and also works with irregular series.

ts_shift(x, by = NULL)

Arguments

x
ts-boxable time series, an object of class ts, xts, data.frame, data.table, or tibble.

by
passed on to base::seq.Date(). Argument can be specified in several ways: - A number, taken to be in days. - A object of class difftime - A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".

Value

a ts-boxable time series, with the same class as the input. If time stamp shifting causes the object to be irregular, a data frame is returned.

See also

ts_lag(), for shifting regular series. date_shift(), for shifting Date vectors.

Examples

ts_shift(AirPassengers, "1 month")

#>      Jan Feb Mar Apr May Jun Jul Aug Sep Oct Nov Dec
#> 1949     112 118 132 129 121 135 148 148 136 119 104
#> 1950 118 115 126 141 135 125 149 170 170 158 133 114
#> 1951 140 145 150 178 163 172 178 199 199 184 162 146
#> 1952 166 171 180 193 181 183 218 230 242 209 191 172
#> 1953 194 196 196 236 235 229 243 264 272 237 211 180
#> 1954 201 204 188 235 227 234 264 302 293 259 229 203
#> 1955 229 242 233 267 269 270 315 364 347 312 274 237
#> 1956 278 284 277 317 313 318 374 413 405 355 306 271
#> 1957 306 315 301 356 348 355 422 465 467 404 347 305
#> 1958 336 340 318 362 348 363 435 491 505 404 359 310
#> 1959 337 360 342 406 396 420 472 548 559 463 407 362
#> 1960 405 417 391 419 461 472 535 622 606 508 461 390
#> 1961 432                                            

ts_shift(AirPassengers, 3)

#> time column needs to specified as the first date of the period, returning data.frame

#>           time value
#> 1   1949-01-04   112
#> 2   1949-02-04   118
#> 3   1949-03-04   132
#> 4   1949-04-04   129
#> 5   1949-05-04   121
#> 6   1949-06-04   135
#> 7   1949-07-04   148
#> 8   1949-08-04   148
#> 9   1949-09-04   136
#> 10  1949-10-04   119
#> 11  1949-11-04   104
#> 12  1949-12-04   118
#> 13  1950-01-04   115
#> 14  1950-02-04   126
#> 15  1950-03-04   141
#> 16  1950-04-04   135
#> 17  1950-05-04   125
#> 18  1950-06-04   149
#> 19  1950-07-04   170
#> 20  1950-08-04   170
#> 21  1950-09-04   158
#> 22  1950-10-04   133
#> 23  1950-11-04   114
#> 24  1950-12-04   140
#> 25  1951-01-04   145
#> 26  1951-02-04   150
#> 27  1951-03-04   178
#> 28  1951-04-04   163
#> 29  1951-05-04   172
#> 30  1951-06-04   178
#> 31  1951-07-04   199
#> 32  1951-08-04   199
#> 33  1951-09-04   184
#> 34  1951-10-04   162
#> 35  1951-11-04   146
#> 36  1951-12-04   166
#> 37  1952-01-04   171
#> 38  1952-02-04   180
#> 39  1952-03-04   193
#> 40  1952-04-04   181
#> 41  1952-05-04   183
#> 42  1952-06-04   218
#> 43  1952-07-04   230
#> 44  1952-08-04   242
#> 45  1952-09-04   209
#> 46  1952-10-04   191
#> 47  1952-11-04   172
#> 48  1952-12-04   194
#> 49  1953-01-04   196
#> 50  1953-02-04   196
#> 51  1953-03-04   236
#> 52  1953-04-04   235
#> 53  1953-05-04   229
#> 54  1953-06-04   243
#> 55  1953-07-04   264
#> 56  1953-08-04   272
#> 57  1953-09-04   237
#> 58  1953-10-04   211
#> 59  1953-11-04   180
#> 60  1953-12-04   201
#> 61  1954-01-04   204
#> 62  1954-02-04   188
#> 63  1954-03-04   235
#> 64  1954-04-04   227
#> 65  1954-05-04   234
#> 66  1954-06-04   264
#> 67  1954-07-04   302
#> 68  1954-08-04   293
#> 69  1954-09-04   259
#> 70  1954-10-04   229
#> 71  1954-11-04   203
#> 72  1954-12-04   229
#> 73  1955-01-04   242
#> 74  1955-02-04   233
#> 75  1955-03-04   267
#> 76  1955-04-04   269
#> 77  1955-05-04   270
#> 78  1955-06-04   315
#> 79  1955-07-04   364
#> 80  1955-08-04   347
#> 81  1955-09-04   312
#> 82  1955-10-04   274
#> 83  1955-11-04   237
#> 84  1955-12-04   278
#> 85  1956-01-04   284
#> 86  1956-02-04   277
#> 87  1956-03-04   317
#> 88  1956-04-04   313
#> 89  1956-05-04   318
#> 90  1956-06-04   374
#> 91  1956-07-04   413
#> 92  1956-08-04   405
#> 93  1956-09-04   355
#> 94  1956-10-04   306
#> 95  1956-11-04   271
#> 96  1956-12-04   306
#> 97  1957-01-04   315
#> 98  1957-02-04   301
#> 99  1957-03-04   356
#> 100 1957-04-04   348
#> 101 1957-05-04   355
#> 102 1957-06-04   422
#> 103 1957-07-04   465
#> 104 1957-08-04   467
#> 105 1957-09-04   404
#> 106 1957-10-04   347
#> 107 1957-11-04   305
#> 108 1957-12-04   336
#> 109 1958-01-04   340
#> 110 1958-02-04   318
#> 111 1958-03-04   362
#> 112 1958-04-04   348
#> 113 1958-05-04   363
#> 114 1958-06-04   435
#> 115 1958-07-04   491
#> 116 1958-08-04   505
#> 117 1958-09-04   404
#> 118 1958-10-04   359
#> 119 1958-11-04   310
#> 120 1958-12-04   337
#> 121 1959-01-04   360
#> 122 1959-02-04   342
#> 123 1959-03-04   406
#> 124 1959-04-04   396
#> 125 1959-05-04   420
#> 126 1959-06-04   472
#> 127 1959-07-04   548
#> 128 1959-08-04   559
#> 129 1959-09-04   463
#> 130 1959-10-04   407
#> 131 1959-11-04   362
#> 132 1959-12-04   405
#> 133 1960-01-04   417
#> 134 1960-02-04   391
#> 135 1960-03-04   419
#> 136 1960-04-04   461
#> 137 1960-05-04   472
#> 138 1960-06-04   535
#> 139 1960-07-04   622
#> 140 1960-08-04   606
#> 141 1960-09-04   508
#> 142 1960-10-04   461
#> 143 1960-11-04   390
#> 144 1960-12-04   432