Time Windows

Filter series for a time range.

ts_window(x, start = NULL, end = NULL)

ts_start(x)

ts_end(x)

Arguments

x
ts-boxable time series, an object of class ts, xts, data.frame, data.table, or tibble.

start
start date, character string, Date or POSIXct

end
end date, character string, Date or POSIXct.

Value

a ts-boxable time series, with the same class as the input.

Examples

# It's fine to use an end date outside of series span
ts_window(mdeaths, end = "2001-01-01")

#>       Jan  Feb  Mar  Apr  May  Jun  Jul  Aug  Sep  Oct  Nov  Dec
#> 1974 2134 1863 1877 1877 1492 1249 1280 1131 1209 1492 1621 1846
#> 1975 2103 2137 2153 1833 1403 1288 1186 1133 1053 1347 1545 2066
#> 1976 2020 2750 2283 1479 1189 1160 1113  970  999 1208 1467 2059
#> 1977 2240 1634 1722 1801 1246 1162 1087 1013  959 1179 1229 1655
#> 1978 2019 2284 1942 1423 1340 1187 1098 1004  970 1140 1110 1812
#> 1979 2263 1820 1846 1531 1215 1075 1056  975  940 1081 1294 1341

ms <- ts_window(mdeaths, end = "1976-12-01")
ts_window(ts_c(fdeaths, ms), start = ts_start(ms), end = ts_end(ms))

#>          fdeaths   ms
#> Jan 1974     901 2134
#> Feb 1974     689 1863
#> Mar 1974     827 1877
#> Apr 1974     677 1877
#> May 1974     522 1492
#> Jun 1974     406 1249
#> Jul 1974     441 1280
#> Aug 1974     393 1131
#> Sep 1974     387 1209
#> Oct 1974     582 1492
#> Nov 1974     578 1621
#> Dec 1974     666 1846
#> Jan 1975     830 2103
#> Feb 1975     752 2137
#> Mar 1975     785 2153
#> Apr 1975     664 1833
#> May 1975     467 1403
#> Jun 1975     438 1288
#> Jul 1975     421 1186
#> Aug 1975     412 1133
#> Sep 1975     343 1053
#> Oct 1975     440 1347
#> Nov 1975     531 1545
#> Dec 1975     771 2066
#> Jan 1976     767 2020
#> Feb 1976    1141 2750
#> Mar 1976     896 2283
#> Apr 1976     532 1479
#> May 1976     447 1189
#> Jun 1976     420 1160
#> Jul 1976     376 1113
#> Aug 1976     330  970
#> Sep 1976     357  999
#> Oct 1976     445 1208
#> Nov 1976     546 1467
#> Dec 1976     764 2059