Omit NA values

Remove NA values in ts-boxable objects, turning explicit into implicit missing values.

ts_na_omit(x)

Arguments

x
ts-boxable time series, an object of class ts, xts, data.frame, data.table, or tibble.

Value

a ts-boxable time series, with the same class as the input.

Details

Note that internal NAs in ts time series will not be removed, as this conflicts with the regular structure.

See also

ts_regular, for the opposite, turning implicit into explicit missing values.

Examples

x <- AirPassengers
x[c(10, 15)] <- NA

# does not change anything for ts objects
ts_na_omit(x)

#>      Jan Feb Mar Apr May Jun Jul Aug Sep Oct Nov Dec
#> 1949 112 118 132 129 121 135 148 148 136  NA 104 118
#> 1950 115 126  NA 135 125 149 170 170 158 133 114 140
#> 1951 145 150 178 163 172 178 199 199 184 162 146 166
#> 1952 171 180 193 181 183 218 230 242 209 191 172 194
#> 1953 196 196 236 235 229 243 264 272 237 211 180 201
#> 1954 204 188 235 227 234 264 302 293 259 229 203 229
#> 1955 242 233 267 269 270 315 364 347 312 274 237 278
#> 1956 284 277 317 313 318 374 413 405 355 306 271 306
#> 1957 315 301 356 348 355 422 465 467 404 347 305 336
#> 1958 340 318 362 348 363 435 491 505 404 359 310 337
#> 1959 360 342 406 396 420 472 548 559 463 407 362 405
#> 1960 417 391 419 461 472 535 622 606 508 461 390 432

# but turns an explicit NA into an implicit for data.frames
ts_na_omit(ts_dts(x))

#>            time value
#>   1: 1949-01-01   112
#>   2: 1949-02-01   118
#>   3: 1949-03-01   132
#>   4: 1949-04-01   129
#>   5: 1949-05-01   121
#>  ---                 
#> 138: 1960-08-01   606
#> 139: 1960-09-01   508
#> 140: 1960-10-01   461
#> 141: 1960-11-01   390
#> 142: 1960-12-01   432